+++ src/app/common/components/advanced-search-dialog/advanced-search-dialog.component.ts
@@ -1,6 +1,9 @@
 import { Component, Inject } from '@angular/core';
 import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
-import { MatLegacyDialogRef as MatDialogRef, MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA } from '@angular/material/legacy-dialog';
+import {
+  MatLegacyDialogRef as MatDialogRef,
+  MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA,
+} from '@angular/material/legacy-dialog';
 import { Router } from '@angular/router';
 import { Filters } from 'src/app/types/general.types';
 
@@ -9,21 +12,19 @@ import { Filters } from 'src/app/types/general.types';
   templateUrl: './advanced-search-dialog.component.html',
   styleUrls: ['./advanced-search-dialog.component.scss'],
 })
-
 export class AdvancedSearchDialogComponent {
   advanced_form: UntypedFormGroup;
-  data_source: { title: string, id: string }[] = []; // used in html
 
   constructor(
     public dialogRef: MatDialogRef<AdvancedSearchDialogComponent>,
     @Inject(MAT_DIALOG_DATA) public data: { filters: Filters },
-    private readonly router: Router,
+    private readonly router: Router
   ) {
     this.advanced_form = new UntypedFormGroup({
-      title: new UntypedFormControl(this.data.filters.title),
-      author: new UntypedFormControl(this.data.filters.author),
+      title: new UntypedFormControl(''),
+      author: new UntypedFormControl(''),
       from: new UntypedFormControl(''),
-      to: new UntypedFormControl('')
+      to: new UntypedFormControl(''),
     });
   }